refactor(firefox): introduce FrameManager
Split out frame management from Page into FrameManager.
FrameManager will keep track of page's frames and DOMWorlds.
